+++ src/jvmMain/kotlin/com/bkahlert/kommons/debug/properties.kt
@@ -1,5 +1,9 @@
 package com.bkahlert.kommons.debug
 
+import java.lang.reflect.AccessibleObject
+import java.lang.reflect.InvocationTargetException
+import kotlin.reflect.KCallable
+import kotlin.reflect.KProperty0
 import kotlin.reflect.KProperty1
 import kotlin.reflect.jvm.javaField
 import kotlin.reflect.jvm.javaGetter
@@ -10,26 +14,47 @@ import kotlin.reflect.jvm.javaGetter
  */
 public actual val Any.properties: Map<String, Any?>
     get() = this::class.members
-        .filterIsInstance<KProperty1<Any, *>>()
         .reversed()
-        .mapNotNull { prop ->
-            prop.getSafely(this@properties).fold({ value -> prop.name to value }, { null })
+        .mapNotNull { prop: KCallable<*> ->
+            @Suppress("UNCHECKED_CAST")
+            when (prop) {
+                is KProperty0<*> -> prop.name to prop.getOrElse { "<error:$it>" }
+                is KProperty1<*, *> -> prop.name to (prop as KProperty1<Any, *>).getOrElse(this@properties) { "<error:$it>" }
+                else -> null
+            }
         }.toMap()
 
-internal fun <T, V> KProperty1<T, V>.getSafely(receiver: T): Result<V> {
-    return kotlin.runCatching { get(receiver) }.recover {
-        val getter = this.javaGetter
-        if (getter != null) {
-            if (!getter.accessible) getter.accessible = true
-            @Suppress("UNCHECKED_CAST")
-            return kotlin.runCatching { getter.invoke(receiver) as V }
+/** Gets the value of this property or the result of [onFailure]. */
+public actual fun <V> KProperty0<V>.getOrElse(onFailure: (Throwable) -> V): V =
+    kotlin.runCatching { get() }.getOrElse(onFailure)
+
+/** Gets the value of this property or the result of [onFailure]. */
+public actual fun <T, V> KProperty1<T, V>.getOrElse(receiver: T, onFailure: (Throwable) -> V): V =
+    kotlin.runCatching { get(receiver) }.getOrElse { originalException ->
+        when (val getter = javaGetter) {
+            null -> when (val field = javaField) {
+                null -> onFailure(originalException)
+                else -> field.value(onFailure) { it.get(receiver) }
+            }
+            else -> getter.value(onFailure) { it.invoke(receiver) }
         }
-        val field = this.javaField
-        if (field != null) {
-            if (!field.accessible) field.accessible = true
-            @Suppress("UNCHECKED_CAST")
-            return kotlin.runCatching { field.get(receiver) as V }
-        }
-        return Result.failure(it)
+    }
+
+private inline fun <T : AccessibleObject, V> T.value(
+    noinline onFailure: (Throwable) -> V,
+    get: (T) -> Any?,
+): V {
+    if (!accessible) accessible = true
+    return kotlin.runCatching {
+        val value = get(this)
+        @Suppress("UNCHECKED_CAST")
+        value as V
+    }.getOrElse(handleNativeFailure(onFailure))
+}
+
+private fun <V> handleNativeFailure(onFailure: (Throwable) -> V): (exception: Throwable) -> V = {
+    when (it) {
+        is InvocationTargetException -> onFailure(it.targetException)
+        else -> onFailure(it)
     }
 }